Update DeleteContacts.php
Final Version

+++ LAMP Stack/LAMP Stack/LAMPAPI/DeleteContacts.php
@@ -1,9 +1,6 @@
 <?php
 
 	$inData = getRequestInfo();
-	
-	$searchResults = "";
-	$searchCount = 0;
 
 	$conn = new mysqli("localhost", "TheBeast", "WeLoveCOP4331", "COP4331");
 	if ($conn->connect_error) 
@@ -12,12 +9,24 @@
 	} 
 	else
 	{
-		$stmt = $conn->prepare("DELETE FROM Contacts WHERE ID = ?");
-		$stmt->bind_param("i", $inData["ID"]);
+		$stmt = $conn->prepare("DELETE FROM Contacts WHERE ID = ? and UserId = ?");
+		$stmt->bind_param("ii", $inData["id"], $inData["userId"]);
 		$stmt->execute();
+
+		if ($stmt->affected_rows > 0) 
+			{
+		http_response_code(200);
+		returnWithInfo("Contact deleted successfully");
+		}
+		else
+			{
+		http_response_code(404);
+		returnWithError("Contact not found");
+		}
 		
 		$stmt->close();
 		$conn->close();
+
 	}
 
 	function getRequestInfo()
